+++ server/apps/auth.py
@@ -1,6 +1,6 @@
 from flask import Blueprint, render_template, redirect, url_for, request, flash
 from werkzeug.security import generate_password_hash, check_password_hash
-from models.users_model import Users
+from Models.users_model import Users
 from .. import db
 from flask_login import login_user, login_required, logout_user
