+++ server/apps/mentors.py
@@ -1,6 +1,6 @@
 from server.app import Resource, reqparse, api
-from models.technical_mentor import TechnicalMentor
-from models.config import db
+from Models.technical_mentor import TechnicalMentor
+from Models.config import db
 
 mentor_parser = reqparse.RequestParser()
 mentor_parser.add_argument('username', type=str, required=True, help='Username is required')
