+++ server/apps/student.py
@@ -5,7 +5,7 @@ import os
 from cloudinary.uploader import upload
 from cloudinary.utils import cloudinary_url
 from server.app import app, db
-from models.student_model import Student
+from Models.student_model import Student
 # Configuration for your PostgreSQL database
 # app.config['SQLALCHEMY_DATABASE_URI'] = 'postgresql://username:password@localhost/database_name'
 # db = SQLAlchemy(app)
